data-mining/raw-data/requirements.txt
python-dotenv==0.14.0
google-cloud-storage==1.31.2